+++ src/main/java/com/OnedayOwner/server/platform/reservation/repository/custom/ReservationRepositoryCustom.java
@@ -1,5 +1,6 @@
 package com.OnedayOwner.server.platform.reservation.repository.custom;
 
+import com.OnedayOwner.server.platform.reservation.entity.Reservation;
 import com.querydsl.core.Tuple;
 
 import java.time.LocalDate;
@@ -9,4 +10,7 @@ public interface ReservationRepositoryCustom {
 
     List<Tuple> getMonthlyReservationInfo(
             LocalDate startDate, LocalDate endDate, Long popupId);
+
+    public List<Reservation> findUpcomingReservationsByUserId(Long userId);
+    public List<Reservation> findCompletedReservationsByUserId(Long userId);
 }
