+++ src/main/java/com/OnedayOwner/server/platform/reservation/repository/impl/ReservationRepositoryImpl.java
@@ -1,13 +1,16 @@
 package com.OnedayOwner.server.platform.reservation.repository.impl;
 
+import com.OnedayOwner.server.platform.reservation.entity.Reservation;
 import com.OnedayOwner.server.platform.reservation.repository.custom.ReservationRepositoryCustom;
 import com.querydsl.core.Tuple;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 
+import static com.OnedayOwner.server.platform.popup.entity.QPopupRestaurant.popupRestaurant;
 import static com.OnedayOwner.server.platform.reservation.entity.QReservation.reservation;
 
 public class ReservationRepositoryImpl implements ReservationRepositoryCustom {
@@ -30,5 +33,25 @@ public class ReservationRepositoryImpl implements ReservationRepositoryCustom {
                 .groupBy(reservation.reservationDateTime.dayOfMonth())
                 .fetch();
 
+    }
+
+    // 방문 예정 예약 찾기
+    @Override
+    public List<Reservation> findUpcomingReservationsByUserId(Long userId) {
+        return jpaQueryFactory.selectFrom(reservation)
+                .join(reservation.popupRestaurant, popupRestaurant).fetchJoin()
+                .where(reservation.user.id.eq(userId)
+                        .and(reservation.reservationDateTime.after(LocalDateTime.now())))
+                .fetch();
+    }
+
+    // 방문 완료 예약 찾기
+    @Override
+    public List<Reservation> findCompletedReservationsByUserId(Long userId) {
+        return jpaQueryFactory.selectFrom(reservation)
+                .join(reservation.popupRestaurant, popupRestaurant).fetchJoin()
+                .where(reservation.user.id.eq(userId)
+                        .and(reservation.reservationDateTime.before(LocalDateTime.now())))
+                .fetch();
     }
 }
